🚚 move: supabase api → auth.ts로 이동
